# value_count Aggregation optimization (backport of #54854)