@nik9000 2020-04-10
Hmm - github's [merging](https://github.com/elastic/elasticsearch/commit/b99a50bcb995c8329286d557303c4ddadcbf8c69) this makes it look like I wrote it. Lies! Master has it [right](https://github.com/elastic/elasticsearch/commit/8e8ce967757929b1f71e6a9ce4cbfbf60407867b).